Don't include streaming URL in string output
To make things cleaner, don't assume that the streaming URL should be
part of the string output for a Track. This allows greater flexibility
for rendering the streaming link on a separate line when only the
current track is being shown, or on the same line when multiple lines of
track history are being shown.

Additionally, passing "-last 0" will show the entire track history.