Clean up output; detect location/venue in track names
Use yaml.v2's struct tags to suppress zero-times from being printed.
(Still need to fix this for JSON output.)

Fix edge case in history output where number requested is an integer
base 10 logarithm. E.g., log10(10) = 1, but we need two spaces to show
index for "10", so take the floor of the log10 instead of ceiling and
always add 1.

Add a filter to strip JEMP Radio station breaks from the output when
history or lastN is requested.

A surprise to me, some tracks include a location after the date (inside
the parenthesized date), which the regexp was not accounting for. This
change captures the date in the regexp, but right now it's not stored in
its own field in the type, just rendered into the the title for full
shows. Need to render this into non-full-show titles as well, and
probably store it as its own field in the Track struct.