Fix failing errors
Don't assume the package-level relistenArtists map has been initialized
before the tests run. Manually load the relisten artists map before
tests requiring it are run. This is still clunky, since it depends on
either (1) the presence of a cached artists map in the user's cache
directory, or (2) fetching a fresh map from Relisten (and then writing
it to the cache directory). Instead, a static relisten cache should be
used and loaded from a `testdata` directory. Ideally, it should be
refreshable record if a flag gets passed.

To do: allow Track.String() to render a track string that includes the
streaming URL without depending on a package-level artist map. Update
the Track type to include a reference to an artist map, or, better, a
function that generates a streaming link when provided a track. We can't
directly provide anything like this to the String() function, in order
to maintain compatibility with the Stringer interface.

Also, allow multiple artist filters to be passed.